# docs/swagger.yaml
basePath: /
definitions:
...
  models.Message:
    properties:
      author:
        type: string
      authorId:
        type: integer
      channelId:
        type: integer
      content:
        type: string
      createdAt:
        type: string
      deletedAt:
        description: 撤回时间，nil表示未撤回
        type: string
      fileMeta:
        items:
          type: integer
        type: array
      id:
        type: integer
      mentions:
        description: '提及列表: [{type, id, name?, avatar?}]'
        items:
          type: integer
        type: array
      platform:
        description: '消息来源设备: web, mobile, desktop'
        type: string
      replyTo:
        allOf:
        - $ref: '#/definitions/models.Message'
        description: 关联的被回复消息
      replyToId:
        description: 回复的消息ID
        type: integer
      tempId:
        description: 临时消息ID，用于匹配WebSocket消息
        type: string
      type:
        type: string
    type: object
  models.PinnedMessage:
    properties:
      channelId:
        description: 频道消息时使用
        type: integer
      createdAt:
        type: string
      guildId:
        description: 频道消息时使用
        type: integer
      id:
        type: integer
      messageId:
        type: integer
      pinnedBy:
        description: 谁PIN的
        type: integer
      threadId:
        description: 私聊消息时使用
        type: integer
    type: object
...
info:
  contact: {}
  description: Simple chat server (Gin + Gorm)
  title: Bubble Chat API
  version: 0.1.0
paths:
...
  /api/channels/{id}/pins:
    get:
      parameters:
      - description: Channel ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.PinnedMessage'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List pinned messages
      tags:
      - pins
    post:
      consumes:
      - application/json
      parameters:
      - description: Channel ID
        in: path
        name: id
        required: true
        type: integer
      - description: '{messageId}'
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PinnedMessage'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Pin message
      tags:
      - pins
  /api/channels/{id}/pins/{messageId}:
    delete:
      parameters:
      - description: Channel ID
        in: path
        name: id
        required: true
        type: integer
      - description: Message ID
        in: path
        name: messageId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Unpin message
      tags:
      - pins
...
  /api/dm/threads/{id}/pins:
    get:
      consumes:
      - application/json
      description: 列出私聊中的置顶消息
      parameters:
      - description: Thread ID
        in: path
        name: id
        required: true
        type: integer
      - description: DM Thread ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Message'
            type: array
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 未认证
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 权限不足
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 会话不存在
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - BearerAuth: []
      summary: List pinned DM messages
      tags:
      - pins
      - pins
    post:
      consumes:
      - application/json
      - application/json
      description: 置顶私聊消息
      parameters:
      - description: Thread ID
        in: path
        name: id
        required: true
        type: integer
      - description: '{messageId}'
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: integer
          type: object
      - description: DM Thread ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 未认证
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 权限不足
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 消息不存在
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - BearerAuth: []
      summary: Pin DM message
      tags:
      - pins
      - pins
  /api/dm/threads/{id}/pins/{messageId}:
    delete:
      consumes:
      - application/json
      description: 取消置顶私聊消息
      parameters:
      - description: Thread ID
        in: path
        name: id
        required: true
        type: integer
      - description: Message ID
        in: path
        name: messageId
        required: true
        type: integer
      - description: DM Thread ID
        in: path
        name: id
        required: true
        type: integer
      - description: Message ID
        in: path
        name: messageId
        required: true
        type: integer
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 未认证
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 权限不足
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 消息不存在
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - BearerAuth: []
      summary: Unpin DM message
      tags:
      - pins
      - pins
    put:
      consumes:
      - application/json
      - application/json
      description: 置顶私聊消息
      parameters:
      - description: Thread ID
        in: path
        name: id
        required: true
        type: integer
      - description: '{messageId}'
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: integer
          type: object
      - description: DM Thread ID
        in: path
        name: id
        required: true
        type: integer
      - description: Message ID
        in: path
        name: messageId
        required: true
        type: integer
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 未认证
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 权限不足
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 消息不存在
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - BearerAuth: []
      summary: Pin DM message
      tags:
      - pins
      - pins
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"